My Calculator
Here is a simple calculator developed in JavaScript that I made just for fun and to put into practice some of what I've already learned. The calculator has a slightly exaggerated interface and a meme-inspired background, but it functions like a regular calculator, allowing users to perform basic mathematical operations such as addition, subtraction, multiplication, and division.

Additionally, it features a 'C' button that clears the display. Users can input values, select an operation, and upon clicking the '=' button, the values will be calculated based on the selected operation. However, if the user presses the '=' button while no value is displayed on the screen, a message saying 'Nothing...' will be shown, indicating that no calculation has been performed.

Essentially, it's a simple calculator with a funny visual style that performs basic calculations. If there's anything that can be improved or added, feel free to share your thoughts with me.